#592: Refactor collection centre timeslot model out of cc table file
