+++ src/app/admin/collectionCentresTable/CollectionCentresTable.tsx
@@ -1,11 +1,9 @@
 "use client";
 
 import React, { useCallback, useEffect, useState } from "react";
-import { Schema } from "@/databaseUtils";
 import { logErrorReturnLogId } from "@/logger/logger";
 import supabase from "@/supabaseClient";
 import { subscriptionStatusRequiresErrorMessage } from "@/common/subscriptionStatusRequiresErrorMessage";
-import { ErrorSecondaryText, ErrorTextModalFooter } from "@/app/errorStylingandMessages";
 import {
     GridActionsCellItem,
     GridColDef,
@@ -21,64 +19,19 @@ import SaveIcon from "@mui/icons-material/Save";
 import CancelIcon from "@mui/icons-material/Close";
 import EditIcon from "@mui/icons-material/Edit";
 import StyledDataGrid from "@/app/admin/common/StyledDataGrid";
-import { FormGroup, LinearProgress } from "@mui/material";
+import { LinearProgress } from "@mui/material";
 import {
+    CollectionCentresTableRow,
     fetchCollectionCentresForTable,
     InsertCollectionCentreResult,
     insertNewCollectionCentre,
     UpdateCollectionCentreResult,
     updateDbCollectionCentre,
-    updateDbCollectionCentreTimeSlots,
 } from "@/app/admin/collectionCentresTable/CollectionCentreActions";
 import { EditToolbar } from "@/app/admin/collectionCentresTable/CollectionCentresTableToolbar";
 import Button from "@mui/material/Button";
-import Icon from "@/components/Icons/Icon";
-import { faShoePrints, faTrashAlt } from "@fortawesome/free-solid-svg-icons";
-import {
-    ButtonsDiv,
-    Centerer,
-    ContentDiv,
-    OutsideDiv,
-    SpaceBetween,
-} from "@/components/Modal/ModalFormStyles";
-import Modal from "@/components/Modal/Modal";
-import styled, { useTheme } from "styled-components";
-import { formatDayjsToHoursAndMinutes, formatTimeStringToHoursAndMinutes } from "@/common/format";
-import { DesktopTimePicker } from "@mui/x-date-pickers";
-import dayjs, { Dayjs } from "dayjs";
 import FloatingToast from "@/components/FloatingToast";
-import CheckboxInput from "@/components/DataInput/CheckboxInput";
-import { StyledIcon, StyledIconButton } from "@/components/Icons/IconButton";
-import { Heading } from "@/app/parcels/ActionBar/ActionModals/GeneralActionModal";
-
-export interface CollectionCentresTableRow {
-    acronym: Schema["collection_centres"]["acronym"];
-    name: Schema["collection_centres"]["name"];
-    id: Schema["collection_centres"]["primary_key"];
-    isDelivery: Schema["collection_centres"]["is_delivery"];
-    isShown: Schema["collection_centres"]["is_shown"];
-    timeSlots: Schema["collection_centres"]["time_slots"];
-    isNew: boolean;
-}
-
-export interface FormattedTimeSlot {
-    time: string;
-    isActive: boolean;
-}
-
-export interface FormattedTimeSlotsWithPrimaryKey {
-    primaryKey: Schema["collection_centres"]["primary_key"];
-    timeSlots: FormattedTimeSlot[];
-}
-
-const ModalTimeSlotsContainer = styled.div`
-    max-height: 40vh;
-    overflow-y: auto;
-`;
-
-const ModalTimeSlotRow = styled.div`
-    width: 15rem;
-`;
+import CollectionCentreTimeSlotsModal from "./CollectionCentreTimeSlotsModal";
 
 function getBaseAuditLogForCollectionCentreAction(
     action: string,
@@ -98,46 +51,6 @@ function getBaseAuditLogForCollectionCentreAction(
     };
 }
 
-function getBaseAuditLogForCollectionCentreTimeSlots(
-    action: string,
-    timeSlotsWithPrimaryKey: FormattedTimeSlotsWithPrimaryKey
-): Pick<AuditLog, "action" | "content" | "collectionCentreId"> {
-    const timeSlots = Object.fromEntries(
-        timeSlotsWithPrimaryKey.timeSlots.map((timeSlot) => [timeSlot.time, timeSlot.isActive])
-    );
-    return {
-        action,
-        content: {
-            timeSlots,
-        },
-        collectionCentreId: timeSlotsWithPrimaryKey.primaryKey,
-    };
-}
-
-const formatCollectionCentreTimeSlotDbData = (
-    row: CollectionCentresTableRow
-): FormattedTimeSlotsWithPrimaryKey => {
-    let formattedTimeSlots: FormattedTimeSlot[];
-
-    if (row.timeSlots === null) {
-        formattedTimeSlots = [];
-    } else {
-        formattedTimeSlots = row.timeSlots.map((timeSlot) => {
-            return {
-                time: formatTimeStringToHoursAndMinutes(
-                    timeSlot.time !== null ? timeSlot.time : ""
-                ),
-                isActive: timeSlot.is_active !== null ? timeSlot.is_active : false,
-            };
-        });
-    }
-
-    return {
-        primaryKey: row.id,
-        timeSlots: formattedTimeSlots,
-    };
-};
-
 const CollectionCentresTable: React.FC = () => {
     const [rows, setRows] = useState<CollectionCentresTableRow[]>([]);
     const [rowModesModel, setRowModesModel] = useState<GridRowModesModel>({});
@@ -145,16 +58,8 @@ const CollectionCentresTable: React.FC = () => {
     const [errorMessage, setErrorMessage] = useState<string | null>(null);
     const [existingRowData, setExistingRowData] = useState<CollectionCentresTableRow | null>(null);
     const [timeSlotModalIsOpen, setTimeSlotModalIsOpen] = useState<boolean>(false);
-    const [timeSlotModalErrorMessage, setTimeSlotModalErrorMessage] = useState<string | null>(null);
-    const [timeSlotModalCentreName, setTimeSlotModalCentreName] = useState<string | null>(null);
-    const [timeSlotModalData, setTimeSlotModalData] =
-        useState<FormattedTimeSlotsWithPrimaryKey | null>(null);
-    const [editableIsShown, setEditableIsShown] = useState<boolean>(false);
-    const [collectionTimeSlotValue, setCollectionTimeSlotValue] = useState<Dayjs>();
-    const [addCollectionTimeSlotError, setAddCollectionTimeSlotError] = useState<string | null>(
-        null
-    );
-    const theme = useTheme();
+    const [selectedRowForTimeSlotEdit, setSelectedRowForTimeSlotEdit] =
+        useState<CollectionCentresTableRow | null>(null);
 
     const getCollectionCentresForTable = useCallback(async () => {
         setErrorMessage(null);
@@ -191,119 +96,6 @@ const CollectionCentresTable: React.FC = () => {
             void supabase.removeChannel(subscriptionChannel);
         };
     }, [getCollectionCentresForTable]);
-
-    const handleModalSaveClick = async (): Promise<void> => {
-        if (timeSlotModalData === null) {
-            return;
-        }
-        const { error: updateTimeSlotError } =
-            await updateDbCollectionCentreTimeSlots(timeSlotModalData);
-        const baseAuditLog = getBaseAuditLogForCollectionCentreTimeSlots(
-            "update collection centre time slots",
-            timeSlotModalData
-        );
-
-        if (updateTimeSlotError) {
-            setTimeSlotModalErrorMessage(
-                `Failed to update the collection centre time slots. Log ID: ${updateTimeSlotError.logId}`
-            );
-            await sendAuditLog({
-                ...baseAuditLog,
-                wasSuccess: false,
-                logId: updateTimeSlotError.logId,
-            });
-        }
-
-        await sendAuditLog({ ...baseAuditLog, wasSuccess: true });
-        setTimeSlotModalIsOpen(false);
-    };
-
-    const handleAddSlotClick = async (): Promise<void> => {
-        setEditableIsShown(true);
-        setCollectionTimeSlotValue(dayjs(collectionTimeSlotValue));
-    };
-
-    const checkIfSlotExists = (
-        existingTimeSlotData: FormattedTimeSlotsWithPrimaryKey,
-        newTimeSlot: FormattedTimeSlot
-    ): boolean => {
-        return existingTimeSlotData.timeSlots.some((slot) => slot.time === newTimeSlot.time);
-    };
-
-    const addNewTimeSlotToTimeSlotModalData = (
-        existingTimeSlotData: FormattedTimeSlotsWithPrimaryKey,
-        newTimeSlot: FormattedTimeSlot
-    ): void => {
-        const newTimeSlotArray = [...existingTimeSlotData.timeSlots, newTimeSlot];
-        newTimeSlotArray.sort((slot1, slot2) => slot1.time.localeCompare(slot2.time));
-
-        const updatedTimeSlotModalData: FormattedTimeSlotsWithPrimaryKey = {
-            ...existingTimeSlotData,
-            timeSlots: newTimeSlotArray,
-        };
-
-        setTimeSlotModalData(updatedTimeSlotModalData);
-    };
-
-    const handleSaveSlotClick = async (): Promise<void> => {
-        setAddCollectionTimeSlotError(null);
-        if (timeSlotModalData === null || collectionTimeSlotValue === undefined) {
-            return;
-        }
-        const newTimeSlot: FormattedTimeSlot = {
-            time: formatDayjsToHoursAndMinutes(collectionTimeSlotValue),
-            isActive: true,
-        };
-
-        if (checkIfSlotExists(timeSlotModalData, newTimeSlot)) {
-            setAddCollectionTimeSlotError(
-                "This time slot already exists. Please select a different time."
-            );
-            return;
-        }
-
-        addNewTimeSlotToTimeSlotModalData(timeSlotModalData, newTimeSlot);
-
-        setEditableIsShown(false);
-    };
-
-    const toggleTimeSlotInModalData = (timeLabel: string): void => {
-        if (!timeSlotModalData) {
-            return;
-        }
-
-        const timeSlotIndex = timeSlotModalData.timeSlots.findIndex(
-            (slot) => slot.time === timeLabel
-        );
-        const timeSlot = timeSlotModalData.timeSlots[timeSlotIndex];
-        if (!timeSlot) {
-            return;
-        }
-
-        timeSlot.isActive = !timeSlot.isActive;
-
-        const updatedTimeSlotData: FormattedTimeSlotsWithPrimaryKey = {
-            ...timeSlotModalData,
-            timeSlots: timeSlotModalData.timeSlots,
-        };
-
-        setTimeSlotModalData(updatedTimeSlotData);
-    };
-
-    const deleteTimeSlotFromModalData = (timeLabel: string): void => {
-        if (!timeSlotModalData) {
-            return;
-        }
-
-        const updatedTimeSlotData: FormattedTimeSlotsWithPrimaryKey = {
-            ...timeSlotModalData,
-            timeSlots: timeSlotModalData.timeSlots.filter(
-                (timeSlot) => timeSlot.time !== timeLabel
-            ),
-        };
-
-        setTimeSlotModalData(updatedTimeSlotData);
-    };
 
     const handleSaveClick = (id: GridRowId) => () => {
         setRowModesModel((currentValue) => ({
@@ -466,10 +258,7 @@ const CollectionCentresTable: React.FC = () => {
             renderHeader: (params) => <Header {...params} />,
             renderCell: (params) => {
                 const handleEditCollectionCentreTimeSlot = (): void => {
-                    const row = params.row as CollectionCentresTableRow;
-                    const formattedTimeSlotData = formatCollectionCentreTimeSlotDbData(row);
-                    setTimeSlotModalData(formattedTimeSlotData);
-                    setTimeSlotModalCentreName(row.name);
+                    setSelectedRowForTimeSlotEdit(params.row as CollectionCentresTableRow);
                     setTimeSlotModalIsOpen(true);
                 };
 
@@ -567,104 +356,13 @@ const CollectionCentresTable: React.FC = () => {
                 />
             )}
             {timeSlotModalIsOpen && (
-                <Modal
-                    header={
-                        <>
-                            <Icon icon={faShoePrints} color={theme.primary.largeForeground[2]} />{" "}
-                            Edit Collection Centre Time Slots
-                        </>
-                    }
+                <CollectionCentreTimeSlotsModal
+                    selectedCollectionCentreInfo={selectedRowForTimeSlotEdit}
                     isOpen={timeSlotModalIsOpen}
                     onClose={() => {
                         setTimeSlotModalIsOpen(false);
                     }}
-                    headerId="expandedCollectionCentreTimeSlotsModal"
-                    footer={
-                        <SpaceBetween>
-                            {!editableIsShown && (
-                                <Button onClick={handleAddSlotClick} variant="contained">
-                                    Add a new slot
-                                </Button>
-                            )}
-                            {editableIsShown && (
-                                <>
-                                    <Centerer>
-                                        <DesktopTimePicker
-                                            label="New Collection Slot"
-                                            views={["hours", "minutes"]}
-                                            format="HH:mm"
-                                            value={dayjs(collectionTimeSlotValue)}
-                                            onChange={(value) =>
-                                                value !== null && setCollectionTimeSlotValue(value)
-                                            }
-                                        />
-                                        <Button onClick={handleSaveSlotClick} variant="contained">
-                                            Save slot
-                                        </Button>
-                                    </Centerer>
-                                    {addCollectionTimeSlotError && (
-                                        <ErrorTextModalFooter>
-                                            {addCollectionTimeSlotError}
-                                        </ErrorTextModalFooter>
-                                    )}
-                                </>
-                            )}
-                            <Button onClick={handleModalSaveClick} variant="contained">
-                                Save
-                            </Button>
-                        </SpaceBetween>
-                    }
-                >
-                    <OutsideDiv>
-                        <ContentDiv>
-                            <Centerer>
-                                <Heading>{timeSlotModalCentreName}</Heading>
-                            </Centerer>
-                            <Centerer>
-                                <ModalTimeSlotsContainer>
-                                    <FormGroup>
-                                        {timeSlotModalData &&
-                                            timeSlotModalData.timeSlots.map((timeSlot) => {
-                                                return (
-                                                    <ModalTimeSlotRow key={timeSlot.time}>
-                                                        <SpaceBetween>
-                                                            <CheckboxInput
-                                                                label={timeSlot.time}
-                                                                checked={timeSlot.isActive}
-                                                                onChange={() =>
-                                                                    toggleTimeSlotInModalData(
-                                                                        timeSlot.time
-                                                                    )
-                                                                }
-                                                            />
-                                                            <StyledIconButton
-                                                                onClick={() =>
-                                                                    deleteTimeSlotFromModalData(
-                                                                        timeSlot.time
-                                                                    )
-                                                                }
-                                                                aria-label="delete"
-                                                            >
-                                                                <StyledIcon
-                                                                    icon={faTrashAlt}
-                                                                    onHoverText={`Delete timeslot ${timeSlot.time}`}
-                                                                />
-                                                            </StyledIconButton>
-                                                        </SpaceBetween>
-                                                    </ModalTimeSlotRow>
-                                                );
-                                            })}
-                                    </FormGroup>
-                                </ModalTimeSlotsContainer>
-                            </Centerer>
-                        </ContentDiv>
-                        <ButtonsDiv>
-                            {timeSlotModalErrorMessage && (
-                                <ErrorSecondaryText>{timeSlotModalErrorMessage}</ErrorSecondaryText>
-                            )}
-                        </ButtonsDiv>
-                    </OutsideDiv>
-                </Modal>
+                ></CollectionCentreTimeSlotsModal>
             )}
         </>
     );
